+++ src/lib.rs
@@ -5,6 +5,7 @@ pub(crate) mod chromedriver;
 mod download;
 mod error;
 pub(crate) mod mgr;
+mod output;
 pub(crate) mod port;
 #[cfg(any(feature = "thirtyfour"))]
 pub(crate) mod session;
@@ -14,6 +15,9 @@ pub use chrome_for_testing::Version;
 pub use chromedriver::Chromedriver;
 pub use error::{ChromeForTestingArtifact, ChromeForTestingManagerError};
 pub use mgr::{ChromeForTestingManager, VersionRequest};
+pub use output::{
+    ChromedriverRunConfig, DriverOutputLine, DriverOutputListener, DriverOutputSource,
+};
 pub use port::{Port, PortRequest};
 #[cfg(any(feature = "thirtyfour"))]
 pub use session::Session;