+++ tests/single_session.rs
@@ -7,8 +7,7 @@ mod common;
 async fn single_session() -> Result<(), Report> {
     tracing_subscriber::fmt().try_init().ok();
 
-    // NOTE: Using beta channel as stable channel chromedriver was bugged on Linux...
-    Chromedriver::run_latest_beta()
+    Chromedriver::run(ChromedriverRunConfig::default())
         .await?
         .with_session(common::wikipedia::test_wikipedia)
         .await?;